+++ respyratorgui/kivygui/app.kv
@@ -15,9 +15,9 @@
 # along with Respyrator.  If not, see <http://www.gnu.org/licenses/>.
 
 #:kivy 1.11.1
-#:import colors respyratorgui.kivygui.kivyconfig.colors
-#:import imgs respyratorgui.kivygui.kivyconfig.imgs
-#:import txts respyratorgui.kivygui.kivyconfig.txts
+#:import colors respyratorgui.kivygui.colors
+#:import imgs respyratorgui.kivygui.imgs
+#:import txts respyratorgui.kivygui.txts
 
 <GuiManager>:
     id: gui_manager
